api/requirements.txt
fastapi
uvicorn[standard]
pydantic
psycopg2-binary
pyjwt